Change request/health counters from u32 to u64 to prevent overflow

+++ snakeway-conf/src/types/runtime/service/health_check.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Deserialize, Serialize, Default)]
 pub struct HealthCheckConfig {
     pub enable: bool,
-    pub failure_threshold: u32,
+    pub failure_threshold: u64,
     pub unhealthy_cooldown_seconds: u64,
 }
